Refactor import paths and fix language property in translation functions
